+++ anyblok/bloks/io/tests/test_exporter_formater.py
@@ -71,10 +71,6 @@ class TestExporterFormater(BlokTestCase):
         value = self.get_value("selection", "Selection")
         self.assertEqual(value, "selection")
 
-    def test_small_interger(self):
-        value = self.get_value(10, "SmallInteger")
-        self.assertEqual(value, "10")
-
     def test_string(self):
         value = self.get_value("selection", "String")
         self.assertEqual(value, "selection")