+++ doc/CHANGES.rst
@@ -24,7 +24,7 @@ CHANGELOG
 * [FIX] ``Blok.import_file`` have not raise if does not found 
   **error_found** in the result of the import
 * [FIX] ``Model.get_primary_keys`` with polymorphism (single table)
-* Remove ``column.uText``, ``column.uString``
+* Remove ``column.uText``, ``column.uString``, ``column.SmallInteger``
 
 0.17.2 (2018-02-27)
 -------------------